Refactor TeamsController and TeamsService to improve organization filtering and member retrieval logic
